requirements.txt
requests
matplotlib
pillow
influxdb-client
pytz
tzlocal
